requirements/base.txt
Django==1.6.2
djangorestframework==2.3.12
Markdown==2.3.1
South==0.8.4
django-model-utils==2.0.1
django-cors-headers==0.12
